fix: partially solve memory issues for huge images in morphology
The main problem was caused by int overflow when input images had big dimensions, as 2000x2000x2000. Changing size and index calculations for size_t with proper casting corrected this. Yet, there is still a misterious problem in cudaMemcpy calls, to copy back deviceOutput to hostOutput in the end of the second chunk execution. This happens for some gpuMemory values in different functions. Sometimes smaller or higher gpuMMemory values make everything work, but not the same gpuMemory for all funciton. 0.9 worked for almost everything except for morph_chain_binary, which the highest working value was 0.6, but this value doesn't work for complement_binary.

+++ src/morphology/binary/complement_binary.cu
@@ -52,7 +52,7 @@ void complement_binary_on_device(dtype* hostImage, dtype* hostOutput, const int 
                                  const int ysize, const int zsize, const int flag_verbose) {
 
   // Set input dimension
-  size_t size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * ysize * zsize;
   size_t nBytes = size * sizeof(dtype);
 
   // Malloc device global memory
