+++ src/morphology/binary/logical_or_binary.cu
@@ -52,7 +52,7 @@ void logical_or_on_device(dtype* hostImage, dtype* hostOutput, const int xsize,
                                  const int ysize, const int zsize, const int flag_verbose) {
 
   // Set input dimension
-  size_t size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * ysize * zsize;
   size_t nBytes = size * sizeof(dtype);
 
   // Malloc device global memory
