+++ src/morphology/binary/reconstruction_binary.cu
@@ -13,7 +13,7 @@ void reconstruction_binary(dtype* deviceMarker, dtype* deviceMask, dtype* device
                            const int xsize, const int ysize, const int zsize, MorphOp operation,
                            const int flag_verbose) {
   // set input dimension
-  size_t size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * ysize * zsize;
   size_t nBytes = size * sizeof(dtype);
 
   // Reconstruction: iterate geodesic erosion/dilation until convergency
@@ -71,7 +71,7 @@ void reconstruction_binary_on_device(dtype* hostImage, dtype* hostMask, dtype* h
                                      const int xsize, const int ysize, const int zsize,
                                      MorphOp operation, const int flag_verbose) {
   // set input dimension
-  size_t size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * ysize * zsize;
   size_t nBytes = size * sizeof(dtype);
 
   // malloc device global memory
@@ -130,7 +130,7 @@ void reconstruction_binary_on_host(dtype* hostImage, dtype* hostMask, dtype* hos
   int flagConverged = 0;
 
   // set input dimension
-  size_t size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * ysize * zsize;
   size_t nBytes = size * sizeof(dtype);
 
   // allocate marker memory
