+++ src/morphology/binary/smooth_binary.cu
@@ -13,7 +13,7 @@ void smooth_binary_on_device(dtype* hostImage, dtype* hostOutput, const int xsiz
                              int kernel_zsize) {
 
   // set input dimension
-  size_t size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * ysize * zsize;
   size_t nBytes = size * sizeof(dtype);
   size_t nBytes_padding = xsize * ysize * (padding_bottom + padding_top) * sizeof(dtype);
 
@@ -129,7 +129,7 @@ void smooth_binary_on_host(dtype* hostImage, dtype* hostOutput, const int xsize,
                            int kernel_zsize) {
 
   // set input dimension
-  size_t size = static_cast<size_t>(xsize) * static_cast<size_t>(ysize) * static_cast<size_t>(zsize);
+  size_t size = static_cast<size_t>(xsize) * ysize * zsize;
   size_t nBytes = size * sizeof(dtype);
 
   // allocate temporary memory
