+++ tests_cuda/morphology/binary/test_geodesic_morph_binary.cpp
@@ -35,7 +35,7 @@ void test_geodesic_morph_binary_on_device(const std::string& filename, const int
   printf("\nTest binary geodesic %s on device\n", (operation ? "dilation" : "erosion"));
 
   // Set input dimension
-  size_t size = static_cast<size_t>(xsize) * static_cast<size_t>(ysize) * static_cast<size_t>(zsize);
+  size_t size = static_cast<size_t>(xsize) * ysize * zsize;
   size_t nBytes = size * sizeof(int);
 
   if (flag_verbose)
@@ -114,7 +114,7 @@ void test_geodesic_morph_binary_on_host(const std::string& filename, const int x
   printf("\nTest binary geodesic %s on host\n", (operation ? "dilation" : "erosion"));
 
   // Set input dimension
-  size_t size = static_cast<size_t>(xsize) * static_cast<size_t>(ysize) * static_cast<size_t>(zsize);
+  size_t size = static_cast<size_t>(xsize) * ysize * zsize;
   size_t nBytes = size * sizeof(int);
 
   if (flag_verbose)
